Use temp dir to create and extract archive and validate contents
By creating and extracting the test archive in a temporary directory, we
prevent test artifacts from staying in the repository, if the test
fails. The `tempfile` crate takes care of deleting temp dirs on exit.

Also the test now validates that the contents of the extracted file
match the contents of the initial file.